+++ ibis_substrait/tests/compiler/test_compiler.py
@@ -6,8 +6,8 @@ import pytest
 from google.protobuf import json_format
 
 from ibis_substrait.compiler.translate import translate
-from ibis_substrait.proto.substrait import algebra_pb2 as stalg
-from ibis_substrait.proto.substrait import type_pb2 as stt
+from ibis_substrait.proto.substrait.ibis import algebra_pb2 as stalg
+from ibis_substrait.proto.substrait.ibis import type_pb2 as stt
 
 NULLABILITY_NULLABLE = stt.Type.Nullability.NULLABILITY_NULLABLE
 NULLABILITY_REQUIRED = stt.Type.Nullability.NULLABILITY_REQUIRED
